Refactor/39/user article relationship (#53)
* chore : Sql 로깅레벨

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : long to integer, override toString method

https://github.com/softeerbootcamp/Team3/issues/39

* style : 게시물 Create 기능 메서드 이름 컨벤션에 따라 변경

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : long to integer

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : long to integer

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : long to integer

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : long to integer, override toString method

https://github.com/softeerbootcamp/Team3/issues/39

* feat : 게스트로 참여하는 글 조회, 호스팅하는 모임 조회 기능 구현

https://github.com/softeerbootcamp/Team3/issues/39

* refactor : Long to Integer

https://github.com/softeerbootcamp/Team3/issues/39

* feat : User - Article N:M Mapping 레포지토리 생성

https://github.com/softeerbootcamp/Team3/issues/39

* test : User - Article N:M Mapping 테스트

https://github.com/softeerbootcamp/Team3/issues/39

* feat : 게스트 혹은 호스트로 참여하는 글 조회 기능 구현

https://github.com/softeerbootcamp/Team3/issues/39

+++ backend/gathergo/src/test/java/lightning/gathergo/repository/ArticleRepositoryTest.java
@@ -38,7 +38,7 @@ class ArticleRepositoryTest {
 
         //when
         repo.save(article);
-        Long storedId = repo.getLastInsertedId();
+        Integer storedId = repo.getLastInsertedId();
 
         //then
         Assertions.assertThat(repo.findById(storedId).get().getTitle())
@@ -74,7 +74,7 @@ class ArticleRepositoryTest {
         // given
         Article article = new Article();
         article.setTitle("변경 전");
-        Long id = repo.save(article).getId();
+        Integer id = repo.save(article).getId();
 
         //when
         repo.updateArticleById("변경 후", null, 0, 0, false, null, null, null, 0, 0, id);
@@ -91,7 +91,7 @@ class ArticleRepositoryTest {
         Article nullArticle = new Article();
         article.setTitle("dummy");
         repo.save(article);
-        Long dummyId = repo.getLastInsertedId();
+        Integer dummyId = repo.getLastInsertedId();
 
         // when
         repo.deleteById(dummyId);
